requirements.txt
ckancrawler
libhxl>=4.25.1
